requirements.txt
# Core dependencies
pandas>=1.5.3
numpy>=1.24.3
python-dotenv>=1.0.0
requests>=2.30.0
aiohttp>=3.8.5
pydantic>=2.0.0
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
streamlit>=1.27.0
plotly>=5.15.0
psutil>=5.9.5

# Web3 and blockchain dependencies
web3>=6.0.0
eth-brownie>=1.19.3
eth-account>=0.8.0
etherscan-python>=2.1.0

# Browser automation
browser-use>=0.2.8
playwright>=1.38.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.7.0
pylint>=2.17.0
pre-commit>=3.3.3

# Database and data handling
SQLAlchemy>=2.0.0
polars>=0.17.11
pyarrow>=12.0.0

# Machine learning
scikit-learn==1.4.0
matplotlib==3.8.2
joblib==1.3.2

# Dashboard & visualization
pydeck==0.8.0
watchdog==3.0.0

# Debugging and profiling
rich==13.7.0
debugpy==1.8.0

# Development
isort==5.13.2
flake8==6.1.0
mypy==1.8.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Async support
asyncio==3.4.3

# Utilities
python-dateutil==2.8.2

# For future expansion - natural language processing
nltk==3.8.1
sentence-transformers==2.5.1 